Test existence capteurs

+++ controller/user.php
@@ -163,16 +163,21 @@ switch ($page) {
             $num = htmlspecialchars($_POST['numeroDeSerie']);
             $numCemac = htmlspecialchars($_POST['Cemac']);
             $nomCapteur = htmlspecialchars($_POST['nomCapteur']);
-            addProduct($num, $_POST['idPiece'], $_SESSION['userID'], $numCemac, $nomCapteur);
-            if(getActionneurModele($_POST['numeroDeSerie']) == 'a'){
-                addActuator($num, $numCemac, $_SESSION['userID']);
+            if (empty(existenceActionneurs($_POST['numeroDeSerie'])) AND empty(existenceCapteurs($_POST['numeroDeSerie']))){
+                addProduct($num, $_POST['idPiece'], $_SESSION['userID'], $numCemac, $nomCapteur);
+                if(getActionneurModele($_POST['numeroDeSerie']) == 'a'){
+                    addActuator($num, $numCemac, $_SESSION['userID']);
+                    
+                } else{
+                    addSensor($num, $numCemac);
+                    
+                }
+                echo "<script type='text/javascript'>document.location.replace('index.php?target=user&page=logements');</script>";
+                exit();
             } else{
-                addSensor($num, $numCemac);
+                $alerte = 'Vous ne pouvez pas ajouter ce produit car il a déjà été enregistré !';
             }
-            
-            echo "<script type='text/javascript'>document.location.replace('index.php?target=user&page=logements');</script>";
-            exit();
-            
+     
         }
         
         
@@ -203,9 +208,7 @@ switch ($page) {
                 exit();
                 
             } else{
-                echo '<script>alert("Vous devez supprimer les capteurs et actionneurs associés à la Cemac avant de la retirer !");</script>';
-                echo "<script type='text/javascript'>document.location.replace('index.php?target=user&page=logements');</script>";
-                exit();
+                $alerte = 'Vous devez supprimer les capteurs et actionneurs associés à la Cemac avant de la retirer !'; 
             }
         }
         if (isset($_POST['number']) AND !empty($_POST['number']) AND isset($_POST['idHouse'])) {
@@ -215,9 +218,7 @@ switch ($page) {
                 echo "<script type='text/javascript'>document.location.replace('index.php?target=user&page=logements');</script>";
                 exit();
             } else{
-                echo '<script>alert("Vous ne pouvez pas ajouter cette Cemac !");</script>';
-                echo "<script type='text/javascript'>document.location.replace('index.php?target=user&page=logements');</script>";
-                exit();
+                $alerte = 'Vous ne pouvez pas ajouter cette Cemac car elle est déjà enregistrée!';
             }
         }
